deduplicate code for roundabout handling
